moved everything from apiCalls.tsx to trackerApiCalls.tsx
