+++ src/pages/JobApplications/JobApplications.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { fetchApplicationsData } from '../../apiCalls';
+import { fetchApplicationsData } from '../../trackerApiCalls';
 import ClipLoader from "react-spinners/ClipLoader";
 import { useUserLoggedContext } from '../../context/UserLoggedContext';
 import { JobApplication } from '../../Interfaces';
